Change filenames
Change filenames to reduce their length.
