+++ src/PANS/Core.hpp
