+++ src/PANS/UISystem.hpp
@@ -1,6 +1,8 @@
-//check if ui system module is active
-#ifdef PANS_MODULE_UISYSTEM
-#include "src/PANS/PANS_UISystem.hpp"
+#pragma once
+#include "Core.hpp"
+
+//module define
+#define PANS_MODULE_UISYSTEM
 
 namespace PANS
 {
@@ -11,46 +13,30 @@ namespace PANS
     {
     public:
       //Destroy this screen - this will invoke destroy on all children
-      virtual void Destroy()
-      {
-
-      }
+      virtual void Destroy() const;
       //setup this screen - this will invoke setup on all children
-      virtual void Setup()
-      {
-
-      }
+      virtual void Setup() const;
     };
-    class Screen : UINode
+    class Screen:public UINode
     {
     public:
       //child elements of the screen
       std::vector<Element> children;
-      //destroy this screen
-      void Destroy()
-      {
-
-      }
-      //setup this screen
-      void Setup()
-      {
-
-      }
     };
-    class Element : UINode
+    class Element:public UINode
     {
     public:
       //position
       int xPos;
       int yPos;
     };
-    class ConfigWindow : Screen
+    class ConfigWindow:public Screen
     {
     public:
       //title for the config screen
       std::string title;
     };
-    class Button : Element
+    class Button:public Element
     {
     public:
       //title for the button
@@ -66,4 +52,3 @@ namespace PANS
     };
   }
 }
-#endif
