+++ src/opcontrol.cpp
@@ -1,4 +1,5 @@
-
+#include "PANS/Core.hpp"
+#include "PANS/UISystem.hpp"
 
 void opcontrol()
 {
